Consolidate desktop webpack config
