+++ apps/desktop/package.json
@@ -21,18 +21,18 @@
     "build-native": "cd desktop_native && node build.js",
     "build": "concurrently -n Main,Rend,Prel -c yellow,cyan \"npm run build:main\" \"npm run build:renderer\" \"npm run build:preload\"",
     "build:dev": "concurrently -n Main,Rend,Prel -c yellow,cyan \"npm run build:main:dev\" \"npm run build:renderer:dev\" \"npm run build:preload:dev\"",
-    "build:preload": "cross-env NODE_ENV=production webpack --config webpack.preload.js",
-    "build:preload:dev": "cross-env NODE_ENV=development webpack --config webpack.preload.js",
-    "build:preload:watch": "cross-env NODE_ENV=development webpack --config webpack.preload.js --watch",
+    "build:preload": "cross-env NODE_ENV=production webpack --config webpack.config.js --config-name preload",
+    "build:preload:dev": "cross-env NODE_ENV=development webpack --config webpack.config.js --config-name preload",
+    "build:preload:watch": "cross-env NODE_ENV=development webpack --config webpack.config.js --config-name preload --watch",
     "build:macos-extension:mac": "./desktop_native/macos_provider/build.sh && node scripts/build-macos-extension.js mac",
     "build:macos-extension:mas": "./desktop_native/macos_provider/build.sh && node scripts/build-macos-extension.js mas",
     "build:macos-extension:masdev": "./desktop_native/macos_provider/build.sh && node scripts/build-macos-extension.js mas-dev",
-    "build:main": "cross-env NODE_ENV=production webpack --config webpack.main.js",
-    "build:main:dev": "npm run build-native && cross-env NODE_ENV=development webpack --config webpack.main.js",
-    "build:main:watch": "npm run build-native && cross-env NODE_ENV=development webpack --config webpack.main.js --watch",
-    "build:renderer": "cross-env NODE_ENV=production webpack --config webpack.renderer.js",
-    "build:renderer:dev": "cross-env NODE_ENV=development webpack --config webpack.renderer.js",
-    "build:renderer:watch": "cross-env NODE_ENV=development webpack --config webpack.renderer.js --watch",
+    "build:main": "cross-env NODE_ENV=production webpack --config webpack.config.js --config-name main",
+    "build:main:dev": "npm run build-native && cross-env NODE_ENV=development webpack --config webpack.config.js --config-name main",
+    "build:main:watch": "npm run build-native && cross-env NODE_ENV=development webpack --config webpack.config.js --config-name main --watch",
+    "build:renderer": "cross-env NODE_ENV=production webpack --config webpack.config.js --config-name renderer",
+    "build:renderer:dev": "cross-env NODE_ENV=development webpack --config webpack.config.js --config-name renderer",
+    "build:renderer:watch": "cross-env NODE_ENV=development webpack --config webpack.config.js --config-name renderer --watch",
     "electron": "node ./scripts/start.js",
     "electron:ignore": "node ./scripts/start.js --ignore-certificate-errors",
     "clean:dist": "rimraf ./dist",
